# Add vectorized c code