This PR speeds up c implementation with SIMD.
The compiler of c-lang has the feature of auto-vectorization.

The result is below.  The program of `v3-native` uses SIMD.

```
$ grep "model name" /proc/cpuinfo | head -1
model name      : Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz
$ gcc --version
gcc (Ubuntu 9.3.0-17ubuntu1~20.04) 9.3.0
Copyright (C) 2019 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
$ clang --version
clang version 10.0.0-4ubuntu1
Target: x86_64-pc-linux-gnu
Thread model: posix
InstalledDir: /usr/bin
$ hyperfine --warmup 2  -m 3 ./v1 ./v2 ./v1-native ./v2-native ./v3-native
Benchmark #1: ./v1
  Time (mean ± σ):      3.523 s ±  0.033 s    [User: 3.522 s, System: 0.001 s]
  Range (min … max):    3.500 s …  3.561 s    3 runs

Benchmark #2: ./v2
  Time (mean ± σ):      2.364 s ±  0.035 s    [User: 2.364 s, System: 0.000 s]
  Range (min … max):    2.332 s …  2.401 s    3 runs

Benchmark #3: ./v1-native
  Time (mean ± σ):      3.532 s ±  0.035 s    [User: 3.532 s, System: 0.000 s]
  Range (min … max):    3.493 s …  3.558 s    3 runs

Benchmark #4: ./v2-native
  Time (mean ± σ):      2.290 s ±  0.022 s    [User: 2.290 s, System: 0.000 s]
  Range (min … max):    2.275 s …  2.315 s    3 runs

Benchmark #5: ./v3-native
  Time (mean ± σ):      1.463 s ±  0.010 s    [User: 1.463 s, System: 0.000 s]
  Range (min … max):    1.456 s …  1.474 s    3 runs

Summary
  './v3-native' ran
    1.57 ± 0.02 times faster than './v2-native'
    1.62 ± 0.03 times faster than './v2'
    2.41 ± 0.03 times faster than './v1'
    2.41 ± 0.03 times faster than './v1-native'
```