@jabbalaci 2021-10-04
Thanks, but I'd like to keep it simple. Same algorithm in every language. If we do vectorization here, then we should do it in other languages too. If it's not supported in a language, then the comparison is not fair. C is already fast enough. At the moment we have similar solutions in every language.