@jabbalaci 2021-10-05
You closed it too early. We could create a `_vectorized` folder and put it there. But I don't want to mix it with non-vectorized examples.