@junjihashimoto 2021-10-05
BTW, here is the faster one.
https://gist.github.com/junjihashimoto/fe75780dba77ef3e33e8be605b3cf0c5