@junjihashimoto 2021-10-09
@jabbalaci 
There are two causes of slowdown.

The first is the width of the register,
Do you use `Intel Core i5-2500 CPU @ 3.30GHz`?
It supports avx not avx2. The width of integer-register is 128bit.
https://ark.intel.com/content/www/us/en/ark/products/52209/intel-core-i5-2500-processor-6m-cache-up-to-3-70-ghz.html

My processor is ` Intel(R) Core(TM) i7-10750H CPU @ 2.60GHz`.
It supports avx2. The width of integer-register is 256bit.
https://ark.intel.com/content/www/us/en/ark/products/201837/intel-core-i7-10750h-processor-12m-cache-up-to-5-00-ghz.html

Second one is that the clock speed is slowing down.
Older generations' clock becomes slow due to avx instructions.
https://stackoverflow.com/questions/56852812/simd-instructions-lowering-cpu-frequency