@jabbalaci 2021-10-09
> The width of integer-register is 128bit.

Wait. I always thought that the register size in a 64-bit machine is 64 bits. Was I wrong the whole time?