@jabbalaci 2022-07-28
It is also benchmarked now: https://github.com/jabbalaci/SpeedTests/tree/master/_vectorized